refactor down to 1 component
