--- src/lib/holocene/orderable-list/orderable-container.svelte
@@ -1,9 +1,0 @@
-<ol class="orderable-list {$$props.class}">
-  <slot />
-</ol>
-
-<style lang="postcss">
-  .orderable-list {
-    @apply border-3 border-primary rounded-lg;
-  }
-</style>
